+++ startfortest/predefined/irods/__init__.py
@@ -1,2 +1,0 @@
-from startfortest.predefined.irods.executables import IrodsExecutablesController
-from startfortest.predefined.irods.services import Irods4_1_10ServiceController, IrodsServiceController