+++ useintest/predefined/irods/__init__.py
@@ -1,0 +1,2 @@
+from useintest.predefined.irods.executables import IrodsExecutablesController
+from useintest.predefined.irods.services import Irods4_1_10ServiceController, IrodsServiceController