+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Switched to using native iRODS controllers, executables and helpers (no longer requires `test-with-irods`).
 - Groups predefined controllers by product (e.g. iRODS) not by type (e.g. service).
 - Allows quicker testing with latest versions of controllers only.
+- Rebranded from `startfortest` to `useintest`.
 
 ## 1.0.0
 ### Added